- Fixed bug in multiple_select_or_404 - Moved ItemPurchaseStatusChanger from classes to item_purchase_completer - Updated TransferHistory to work correctly with two generic keys, added TransferHistoryManager to simplify creation of new TransferHistory records - Updated RefundSerializer and ItemPurchaseData with fields user_uuid_from and user_uuid_to - ItemPurchaseUpdaterViewSet renamed to ItemPurchaseUpdateView - Fixed fixtures for TransferHistory
