+++ backend/payments/apps/base/fixtures/test/transfer_histories.json
@@ -3,10 +3,12 @@
         "model": "item_purchases.transferhistory",
         "pk": 1,
         "fields": {
-            "account_type_id": 7,
+            "account_to_type_id": 7,
+            "account_from_type_id": 8,
             "amount": "750.00",
             "amount_currency": "RUB",
-            "object_id": "1",
+            "object_id_from": "1",
+            "object_id_to": "1",
             "created_date": "2023-05-25 22:49:21.682908+00"
         }
     },
@@ -14,10 +16,12 @@
         "model": "item_purchases.transferhistory",
         "pk": 2,
         "fields": {
-            "account_type_id": 8,
+            "account_to_type_id": 8,
+            "account_from_type_id": 8,
             "amount": "250.00",
             "amount_currency": "RUB",
-            "object_id": "1",
+            "object_id_from": "1",
+            "object_id_to": "1",
             "created_date": "2023-05-25 22:49:21.682908+00"
         }
     },
@@ -25,10 +29,12 @@
         "model": "item_purchases.transferhistory",
         "pk": 3,
         "fields": {
-            "account_type_id": 7,
+            "account_to_type_id": 7,
+            "account_from_type_id": 8,
             "amount": "75.00",
             "amount_currency": "RUB",
-            "object_id": "2",
+            "object_id_from": "1",
+            "object_id_to": "2",
             "created_date": "2023-05-26 22:49:21.682908+00"
         }
     },
@@ -36,10 +42,12 @@
         "model": "item_purchases.transferhistory",
         "pk": 4,
         "fields": {
-            "account_type_id": 8,
+            "account_to_type_id": 8,
+            "account_from_type_id": 8,
             "amount": "25.00",
             "amount_currency": "RUB",
-            "object_id": "1",
+            "object_id_from": "1",
+            "object_id_to": "1",
             "created_date": "2023-05-26 22:49:21.682908+00"
         }
     },
@@ -47,10 +55,12 @@
         "model": "item_purchases.transferhistory",
         "pk": 5,
         "fields": {
-            "account_type_id": 7,
+            "account_to_type_id": 7,
+            "account_from_type_id": 8,
             "amount": "7.50",
             "amount_currency": "RUB",
-            "object_id": "2",
+            "object_id_from": "1",
+            "object_id_to": "2",
             "created_date": "2023-05-27 22:49:21.682908+00"
         }
     },
@@ -58,10 +68,12 @@
         "model": "item_purchases.transferhistory",
         "pk": 6,
         "fields": {
-            "account_type_id": 8,
+            "account_to_type_id": 8,
+            "account_from_type_id": 8,
             "amount": "2.50",
             "amount_currency": "RUB",
-            "object_id": "1",
+            "object_id_from": "1",
+            "object_id_to": "1",
             "created_date": "2023-05-27 22:49:21.682908+00"
         }
     }
