+++ backend/payments/apps/base/utils/db_query.py
@@ -11,7 +11,7 @@ def multiple_select_or_404(
     field_name = f'{look_up_field}__in'
     query = list(django_model.objects.filter(**{field_name: data_list}))
 
-    if len(query) != len(data_list):
+    if len(query) != len(set(data_list)):
         missing_attrs = set(data_list) - {getattr(obj, look_up_field) for obj in query}
         raise Http404(
             f'Field {look_up_field} with next values {missing_attrs} not found in database.',
